# api/openapi.yaml
openapi: 3.0.3
info:
  title: CCSP API
  version: 0.1.0
  description: MVP API for Civic Candidate Support Platform.
servers:
  - url: https://api.example.com
paths:
  /auth/register:
    post:
      summary: Register candidate account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Account created
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
  /candidate/profile:
    get:
      summary: Get candidate profile
      responses:
        '200':
          description: Profile payload
    put:
      summary: Create or update candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CandidateProfile'
      responses:
        '200':
          description: Profile updated
...
  /checklist/{itemId}/status:
    patch:
      summary: Update checklist item status
      parameters:
        - in: path
          name: itemId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Status updated
...
components:
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        fullName:
          type: string
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    CandidateProfile:
      type: object
      required: [officeTitle, jurisdiction]
      properties:
        campaignName:
          type: string
        officeTitle:
          type: string
        jurisdiction:
          type: string
        electionDate:
          type: string
          format: date
        partyAffiliation:
          type: string